Data Utils for Puzzle Access, Puzzle Layer Grid Setup
- added data utils to get to puzzle fields
- setup puzzle layer to take puzzle number and draw grid
- added Game Constants class

+++ FishSet/Utils/GridUtils.h
@@ -17,7 +17,7 @@ typedef struct
 
 // helper function to create a GridCoord
 static inline GridCoord
-GridCoordMake(const NSInteger x, const NSInteger y)
+GridCoordMake(const int x, const int y)
 {
 	GridCoord gridCoord = {x, y};
 	return gridCoord;
